move expectation function to separate file
